refactor(identity): simplify asset lock registration flow
